rename storage endpoint to api endpoint (all storage access is proxied though api). Remove minio dependency. cleanup Elasticsearch mapping.
