+++ static/document-processor/settings_folder.json
@@ -1,18 +1,5 @@
 {
-  "settings": {
-    "analysis": {
-      "analyzer": {
-        "fscrawler_path": {
-          "tokenizer": "fscrawler_path"
-        }
-      },
-      "tokenizer": {
-        "fscrawler_path": {
-          "type": "path_hierarchy"
-        }
-      }
-    }
-  },
+  "settings": {},
   "mappings": {
     "properties" : {
       "real" : {
